server/requirements.txt
numpy>=1.0,<2.00
Pillow==4.2.1
thriftpy==0.3.9
future==0.16.0
netifaces==0.10.6
typing>=3.6,<4.0